Split up workspace snapshot graph unit tests
This commit splits up workspace snapshot graph unit tests into two more
submodules. The files were getting way too big and, selfishly, causing
my editor (RustRover) to slow down significantly.
